Misc NOCOMMIT cleanups
