+++ solr/core/src/test/org/apache/solr/filestore/TestDistribFileStore.java
@@ -84,7 +84,7 @@ public class TestDistribFileStore extends SolrCloudTestCase {
             .configure();
     try {
       byte[] derFile = readFile("cryptokeys/pub_key512.der");
-      uploadKey(derFile, ClusterFileStoreAPI.KEYS_DIR + "/pub_key512.der", cluster);
+      uploadKey(derFile, ClusterFileStore.KEYS_DIR + "/pub_key512.der", cluster);
 
       try {
         postFile(
